сделал взаимодействие с сервером на базе Swing
