+++ src/Client/JFrame_Demo.java
@@ -1,0 +1,7 @@
+package Frame;
+
+public class JFrame_Demo {
+    public static void main(String[] args) {
+        new MyWindow();
+    }
+}
