--- src/Lesson_6/JFrame_Demo.java
@@ -1,7 +1,0 @@
-package JFrame;
-
-public class JFrame_Demo {
-    public static void main(String[] args) {
-        new MyWindow();
-    }
-}
